pyproject.toml
[project]
name = "throttling-sequencer-webapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "strawberry-graphql>=0.270.1",
    "structlog>=25.3.0",
    "svcs>=25.1.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ipython>=9.2.0",
    "pudb>=2025.1",
    "pytest>=8.3.5",
]

[tool.ruff]
target-version="py312"
line-length=120
indent-width=4
